Preparar query para relatório de clientes

+++ database/migrations/2019_02_07_143620_create_clientes_table.php
@@ -19,7 +19,7 @@ class CreateClientesTable extends Migration
             $table->timestamp('dt_born')->nullable();
             $table->string('name_mother')->nullable();
             $table->string('name_father')->nullable();
-            $table->float('sexo')->nullable();
+            $table->integer('sexo')->nullable();
             $table->float('est_civil')->nullable();
             $table->string('cpf');
             $table->string('rg')->nullable();
